refactor(transport): the construction contract moves out of the class file
`scan-file-size` refused the tree: the SEC-008 docstring corrections pushed
`ws-transport-configurable.ts` to 305 lines against the 300 ceiling, and the
ratchet's instruction is split, not baseline.

The seam is the honest one rather than a line count: `IWsTransportConfig` and
its defaults are what a CALLER decides before the server exists — port, retries,
and the admission contract whose doc comments this PR just corrected — while the
class is what the transport does with those decisions. 266 + 50 lines, re-export
unchanged, 37 tests pass.
